python module: OT method, reorganize python code
- put transformation related stuff in separate file
